# optimize parallel algorithm