--- a/src/main/java/com/github/bartimaeusnek/bartworks/common/tileentities/multis/mega/GT_TileEntity_MegaBlastFurnace.java
+++ b/src/main/java/com/github/bartimaeusnek/bartworks/common/tileentities/multis/mega/GT_TileEntity_MegaBlastFurnace.java
@@ -59,6 +59,7 @@ import net.minecraftforge.fluids.FluidStack;
 import java.util.*;
 import java.util.stream.Collectors;
 
+import static com.github.bartimaeusnek.bartworks.util.RecipeFinderForParallel.getMultiOutput;
 import static com.github.bartimaeusnek.bartworks.util.RecipeFinderForParallel.handleParallelRecipe;
 import static com.gtnewhorizon.structurelib.structure.StructureUtility.ofBlockAdder;
 import static com.gtnewhorizon.structurelib.structure.StructureUtility.transpose;
@@ -162,9 +163,9 @@ public class GT_TileEntity_MegaBlastFurnace extends GT_MetaTileEntity_ElectricBl
     }
 
     @SuppressWarnings("rawtypes")
-    public ArrayList TTTunnels = new ArrayList<>();
+    public ArrayList<Object> TTTunnels = new ArrayList<>();
     @SuppressWarnings("rawtypes")
-    public ArrayList TTMultiAmp = new ArrayList<>();
+    public ArrayList<Object> TTMultiAmp = new ArrayList<>();
 
     @Override
     public void loadNBTData(NBTTagCompound aNBT) {
@@ -397,45 +398,10 @@ public class GT_TileEntity_MegaBlastFurnace extends GT_MetaTileEntity_ElectricBl
             int tCurrentPara = handleParallelRecipe(tRecipe, tFluids, tInputs, (int) tMaxPara);
             processed = tCurrentPara;
             found_Recipe = true;
-            if (tRecipe.mOutputs != null) {
-                for (int i = 0; i < tRecipe.mOutputs.length; i++) {
-                    tCurrentPara = processed;
-                    while (tCurrentPara > 0 && tRecipe.getOutput(i) != null) {
-                        int maxSize = tRecipe.getOutput(i).getMaxStackSize();
-                        if (tCurrentPara <= maxSize) {
-                            outputItems.add(GT_Utility.copyAmount(maxSize, tRecipe.getOutput(i)));
-                            tCurrentPara -= maxSize;
-                        }
-                        else {
-                            outputItems.add(GT_Utility.copyAmount(tCurrentPara, tRecipe.getOutput(i)));
-                            tCurrentPara = 0;
-                        }
-                    }
-                }
-            }
-            if (tRecipe.mFluidOutputs != null) {
-                for (int i = 0; i < tRecipe.mFluidOutputs.length; i++) {
-                    while (tRecipe.getFluidOutput(i) != null) {
-                        FluidStack tOutputFluid = new FluidStack(tRecipe.getFluidOutput(i).getFluid(), tRecipe.getFluidOutput(i).amount * processed);
-                        outputFluids.add(tOutputFluid);
-                    }
-                }
-            }
+            Object[] Outputs = getMultiOutput(tRecipe, tCurrentPara);
+            outputFluids = (ArrayList<FluidStack>) Outputs[0];
+            outputItems = (ArrayList<ItemStack>) Outputs[1];
         }
-
-       /* while (this.getStoredInputs().size() > 0 && processed < ConfigHandler.megaMachinesMax) {
-            if (this.mHeatingCapacity >= tRecipe.mSpecialValue && (precutRecipeVoltage * (processed + 1)) < nominalV && tRecipe.isRecipeInputEqual(true, tFluids, tInputs)) {
-                found_Recipe = true;
-                for (int i = 0; i < tRecipe.mOutputs.length; i++) {
-                    outputItems.add(tRecipe.getOutput(i));
-                }
-                for (int i = 0; i < tRecipe.mFluidOutputs.length; i++) {
-                    outputFluids.add(tRecipe.getFluidOutput(i));
-                }
-                ++processed;
-            } else
-                break;
-        }*/
 
         if (found_Recipe) {
             this.mEfficiency = (10000 - (this.getIdealStatus() - this.getRepairStatus()) * 1000);