--- a/src/main/java/com/github/bartimaeusnek/bartworks/common/tileentities/multis/mega/GT_TileEntity_MegaDistillTower.java
+++ b/src/main/java/com/github/bartimaeusnek/bartworks/common/tileentities/multis/mega/GT_TileEntity_MegaDistillTower.java
@@ -44,6 +44,7 @@ import net.minecraftforge.fluids.FluidStack;
 import java.util.ArrayList;
 import java.util.List;
 
+import static com.github.bartimaeusnek.bartworks.util.RecipeFinderForParallel.getMultiOutput;
 import static com.github.bartimaeusnek.bartworks.util.RecipeFinderForParallel.handleParallelRecipe;
 import static com.gtnewhorizon.structurelib.structure.StructureUtility.*;
 import static gregtech.api.enums.GT_Values.V;
@@ -227,10 +228,11 @@ public class GT_TileEntity_MegaDistillTower extends GT_MetaTileEntity_Distillati
         FluidStack[] tFluids = tFluidList.toArray(new FluidStack[0]);
         if (tFluids.length > 0) {
             for (FluidStack tFluid : tFluids) {
-                ArrayList<FluidStack[]> outputFluids = new ArrayList<>();
+                ArrayList<FluidStack> outputFluids = new ArrayList<>();
+                ArrayList<ItemStack> outputItems = new ArrayList<>();
+                Object[] Outputs;
                 int processed = 0;
                 boolean found_Recipe = false;
-                FluidStack[] output;
                 GT_Recipe tRecipe = GT_Recipe.GT_Recipe_Map.sDistillationRecipes.findRecipe(this.getBaseMetaTileEntity(), false, GT_Values.V[tTier], new FluidStack[]{tFluid});
 
                 if (tRecipe != null) {
@@ -238,38 +240,13 @@ public class GT_TileEntity_MegaDistillTower extends GT_MetaTileEntity_Distillati
                     long tMaxPara = Math.min(ConfigHandler.megaMachinesMax, nominalV / tRecipe.mEUt);
                     int tCurrentPara = handleParallelRecipe(tRecipe, new FluidStack[]{tFluid}, null, (int) tMaxPara);
                     processed = tCurrentPara;
-                    if (tRecipe.mFluidOutputs != null) {
-                        output = new FluidStack[tRecipe.mFluidOutputs.length];
-                        for (int i = 0; i < tRecipe.mFluidOutputs.length; i++) {
-                            if (tRecipe.getFluidOutput(i) != null) {
-                                output[i] = new FluidStack(tRecipe.getFluidOutput(i).getFluid(), tRecipe.getFluidOutput(i).amount * tCurrentPara);
-                            }
-                        }
-                        outputFluids.add(output);
-                    }
+                    Outputs = getMultiOutput(tRecipe, tCurrentPara);
+                    outputFluids = (ArrayList<FluidStack>) Outputs[0];
+                    outputItems = (ArrayList<ItemStack>) Outputs[1];
                 }
 
-               /* while (this.getStoredFluids().size() > 0 && processed < ConfigHandler.megaMachinesMax) {
-                    if (tRecipe != null && (tRecipe.mEUt * (processed + 1)) < nominalV && tRecipe.isRecipeInputEqual(true, tFluids)) {
-                        found_Recipe = true;
-                        if (tRecipe.mFluidOutputs.length == 1 && tRecipe.mFluidOutputs[0].amount == 0)
-                            tRecipe.mFluidOutputs[0].amount = tRecipe.mFluidInputs[0].amount;
-                        output = new FluidStack[tRecipe.mFluidOutputs.length];
-                        for (int i = 0; i < output.length; i++) {
-                            output[i] = new FluidStack(tRecipe.mFluidOutputs[i],tRecipe.mFluidOutputs[i].amount);
-                        }
-                        outputFluids.add(output);
-                        ++processed;
-                    } else
-                        break;
-                }*/
                 if (!found_Recipe)
                     continue;
-                /*for (int j = 1; j < outputFluids.size(); j++) {
-                    for (int k = 0; k < outputFluids.get(j).length; k++) {
-                        outputFluids.get(0)[k].amount += outputFluids.get(j)[k].amount;
-                    }
-                }*/
                 this.mEfficiency = (10000 - (this.getIdealStatus() - this.getRepairStatus()) * 1000);
                 this.mEfficiencyIncrease = 10000;
                 long actualEUT = (long) (tRecipe.mEUt) * processed;
@@ -289,20 +266,10 @@ public class GT_TileEntity_MegaDistillTower extends GT_MetaTileEntity_Distillati
                     this.mEUt = (-this.mEUt);
                 }
                 this.mMaxProgresstime = Math.max(1, this.mMaxProgresstime);
-                this.mOutputFluids = outputFluids.get(0).clone();
-                if (tRecipe.getOutput(0) != null) {
-                    int stacks = processed / 64;
-                    ItemStack[] outputs = new ItemStack[stacks];
-                    if (stacks > 0) {
-                        for (int i = 0; i < stacks; i++)
-                            if (i != stacks - 1)
-                                outputs[i] = BW_Util.setStackSize(tRecipe.getOutput(0),64);
-                            else
-                                outputs[i] = BW_Util.setStackSize(tRecipe.getOutput(0),processed - (64 * i));
-                        this.mOutputItems = outputs;
-                    } else
-                        this.mOutputItems = null;
-                } else
+                this.mOutputFluids = outputFluids.toArray(new FluidStack[0]);
+                if (!outputItems.isEmpty())
+                    this.mOutputItems = outputItems.toArray(new ItemStack[0]);
+                else
                     this.mOutputItems = null;
                 this.updateSlots();
                 return true;