i did a optimization for the parallel algorithm of mega mechines. directly using division instead of loop to get the parallel recipe amount.
when i tested in MEBF (ticking 256 steel), it uses arv 67,000ns, worst 233,000ns. (the old loop algorithm is arv 180,000ns, worst 310,000ns). it can save more time when the machine have more parallel recipes, especially for mega multis. however, it may perform worse than the origin loop when machine has less 8 parallel recipes.