@bombcar 2021-11-19
The build is failing for the same thing mine local copy is failing on - @SinTh0r4s was working on it but that may have to get merged before the build will succeed: https://discord.com/channels/181078474394566657/603348502637969419/911322404867948584